Moving SearchLayout and PersistentSearchView to their own package
